move span tag logic to useSpanFieldSupportedTags
